requirements.txt
airflow==2.10.5
kafka-python
python-dotenv
pyspark==3.5.0
pandas
